refactor: Update WindowBackgroundFit to use np.float32 instead of np.float16 for p and r arrays

+++ src/hotspot3/fit.py
@@ -91,8 +91,8 @@ class WindowBackgroundFit(BackgroundFit):
         mean, var = self.sliding_mean_and_variance(agg_cutcounts, min_count=self.config.min_mappable_bg)
         p = self.p_from_mean_and_var(mean, var)
         success_fit_mask = ~p.mask
-        p = p.filled(np.nan).astype(np.float16)
-        r = self.r_from_mean_and_var(mean, var).filled(np.nan).astype(np.float32)
+        p = p.filled(np.nan)
+        r = self.r_from_mean_and_var(mean, var).filled(np.nan)
 
         bad_fit = ma.where(mean >= var)[0]
         if len(bad_fit) > 0:
